osutil/user: replace our uses of os/user and filepath.Glob("/home/*")
Previously we were using a mixture of os/user and filepath globs to get at
users in the system. This replaces those uses with an ad-hoc package,
osutil/user, which is initially a very minimal backend that just knows about
files, but is careful to have an interface that should make replacing it with a
bigger, or more modular backend possible down the road with no impact on the
rest of the codebase.

There is still some room for things to break: osutil still includes a cgo-based
implementation of group lookup, which should probably be replaced by something
analogous to osutil/user. AppArmor doesn't afaik really know about user homes
outside of /home/ (except /root?), and this code will produce such users (this
was broken before, in the opposite direction: users with homes outside of
/home/ were partially invisible; in both cases snapd is unusable by those users
without tweaks to the local system's rules).

+++ interfaces/builtin/account_control.go
@@ -89,7 +89,7 @@ func makeAccountControlSecCompSnippet() (string, error) {
 	}
 
 	snippet := strings.Replace(accountControlConnectedPlugSecCompTemplate,
-		"{{group}}", strconv.FormatUint(gid, 10), -1)
+		"{{group}}", strconv.FormatUint(uint64(gid), 10), -1)
 
 	return snippet, nil
 }
